Moving out ResponsiveDrawer and FloatingNodeMenu
